+++ CHANGELOG.md
@@ -32,6 +32,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add an option to ignore statistics before a given date [434](https://github.com/CCSDForge/episciences/issues/434)
   now, except for users, the total number of submissions is the total number of articles published, indicators only include data after the date configured in the journal's settings.
 
+### Added
+- CSL support for data and linked software, addition of management of several formats for the data and linked software form (HAL url, SWH url, arxiv url), extraction of citation from the hal api  if hal is submitted in the form of data and linked software.
+
 ## v1.0.41.7 - 2024-01-23
 ### Changed
 - [COAR Notify] processing of new versions + enhancements