+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 - Add an option to ignore statistics before a given date [434](https://github.com/CCSDForge/episciences/issues/434)
   now, except for users, the total number of submissions is the total number of articles published, indicators only include data after the date configured in the journal's settings.
+### Fixed
+- Submission form validation fails when replacing existing version.
 
 ## v1.0.41.8 - 2024-02-12
 ### Fixed