+++ CHANGELOG.md
@@ -15,7 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Security
 -->
 
-## Unreleased
+## v1.0.41.8 - 2024-02-15
+### Added
 - Add an option to ignore statistics before a given date [434](https://github.com/CCSDForge/episciences/issues/434)
   now, except for users, the total number of submissions is the total number of articles published, indicators only include data after the date configured in the journal's settings.
 ### Fixed